// packages/engine/src/nodes/base/types.ts
export enum NodeType {
	Base = "base-node",
	Group = "group-node",
	Rect = "rect-node",
	Ellipse = "ellipse-node",
	Star = "star-node",
	Polygon = "polygon-node",
	Line = "line-node",
	Text = "text-node",
	Frame = "frame-node",
	Image = "image-node",
	Video = "video-node",
	Path = "path-node",
}

export type Size = {
	width: number;
	height: number;
};

export type Rect = {
	x: number;
	y: number;
	width: number;
	height: number;
};

export type OrientedRect = {
	center: Vector2;
	width: number;
	height: number;
	rotation: number;
};